refactor: add dropdownItem component
