+++ my-dropdown-menu/src/components/dropdown.tsx
@@ -5,42 +5,28 @@ import { ChevronDownIcon, ChevronRightIcon } from '@heroicons/react/20/solid'
 import { SiGithub } from 'react-icons/si';
 import { RxStitchesLogo, RxTwitterLogo } from 'react-icons/rx'
 import { BsCheck } from 'react-icons/bs'
-
-
-function classNames(...classes: (string | boolean)[]): string {
-  return classes.filter(Boolean).join(' ');
-}
+import { DropdownMenuItem, DropdownMenuNestedItem } from './dropdownItem';
 
 // Multi-level dropdown menu
-export function Dropdown() {
+export function DropdownMenu() {
   const [isFavoritesHovered, setIsFavoritesHovered] = useState(false);
   const [focusedElement, setFocusedElement] = useState('favorites');
   const [checked, setChecked] = useState('showToolbar');
 
-  const favoritesRef = React.useRef(null);
+  // const favoritesRef = React.useRef(null);
 
   // TODO: Allow for arrow key navigation
   // Can't go back up to favorites from subMenuItems
   useEffect(() => {
     const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'ArrowDown' && focusedElement === 'window') {
-        setFocusedElement('favorites');
-      }
       if (event.key === 'ArrowRight' && focusedElement === 'favorites') {
         setIsFavoritesHovered(true);
         setFocusedElement('subMenuFirstItem');
-        console.log('event.key', event.key);
-        console.log('focusedElement', focusedElement);
         console.log('isFavoritesHovered', isFavoritesHovered);
       } else if (event.key === 'ArrowDown' && focusedElement === 'favorites') {
         setIsFavoritesHovered(false);
         setFocusedElement('download');
-        console.log('focusedElement', focusedElement);
         console.log('isFavoritesHovered', isFavoritesHovered);
-      }
-
-      if (event.key === 'ArrowDown' && focusedElement === 'subMenuThirdItem') {
-        event.preventDefault();
       }
     };
 
@@ -72,65 +58,18 @@ export function Dropdown() {
         >
           <Menu.Items className="absolute left-1/2 z-10 mt-2 w-56 origin-top-right divide-y divide-gray-100 rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none transform -translate-x-1/2">
             <div className="py-1">
-              <Menu.Item>
-                {({ active }: { active: boolean }) => (
-                  <div
-                    className={classNames(
-                      active ? 'bg-blue-400 text-gray-900' : 'text-gray-700',
-                      'block px-4 py-2 text-sm'
-                    )}
-                  >
-                    <div className='flex flex-row '>
-                      <div className='w-6'>{checked === 'type' ? <BsCheck size="1.5em" color="black" /> : ''} </div>
-                      <div>New Type</div>
-                    </div>
-                  </div>
-                )}
-              </Menu.Item>
-              <Menu.Item>
-                {({ active }: { active: boolean }) => (
-                  <div
-                    role="button"
-                    tabIndex={0}
-                    className={classNames(
-                      active ? 'bg-blue-400 text-gray-900' : 'text-gray-700',
-                      'block px-4 py-2 text-sm'
-                    )}
-                    onFocus={() => { setFocusedElement('window'); }}
-                  >
-                    <div className='flex flex-row '>
-                      <div className='w-6'>{checked === 'window' ? <BsCheck size="1.5em" color="black" /> : ''} </div>
-                      <div>New Windows</div>
-                    </div>
-                  </div>
-                )}
-              </Menu.Item>
+              <DropdownMenuItem label="New Tab" Icon={BsCheck} checked={checked === 'newTab'} />
+              <DropdownMenuItem label='New Window' Icon={BsCheck} checked={checked === 'newWindow'} />
             </div>
             <div className="py-1">
-              <Menu.Item>
-                {({ active }: { active: boolean }) => (
-                  <div
-                    onMouseEnter={() => setIsFavoritesHovered(true)}
-                    onMouseLeave={() => setIsFavoritesHovered(false)}
-                    ref={favoritesRef}
-                    onFocus={() => { setFocusedElement('favorites'); }}
-                  >
-                    <div
-                      className={classNames(
-                        focusedElement === 'favorite' ? 'bg-blue-400 text-gray-900' : active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
-                        'block px-4 py-2 text-sm'
-                      )}
-                    >
-                      <div className="flex flex-row">
-                        <div className='w-6'>{checked === 'favorites' ? <BsCheck size="1.5em" color="black" /> : ''} </div>
-                        <div> Favorites</div>
-                        <div className="justify-self-end"><ChevronRightIcon className="-mr-1 h-5 w-5 text-gray-400" aria-hidden="true" /></div>
-                      </div>
-
-                    </div>
-                  </div>
-                )}
-              </Menu.Item>
+              <div
+                onMouseEnter={() => setIsFavoritesHovered(true)}
+                onMouseLeave={() => setIsFavoritesHovered(false)}
+                // ref={favoritesRef}
+                onFocus={() => { setFocusedElement('favorites'); console.log("onfocused") }}
+              >
+                <DropdownMenuNestedItem label="Favorites" Icon={BsCheck} checked={checked === 'favorites'} ishovered={isFavoritesHovered} />
+              </div>
               {isFavoritesHovered && (
                 <div
                   onMouseEnter={() => setIsFavoritesHovered(true)}
@@ -138,140 +77,19 @@ export function Dropdown() {
                   className="absolute top-20 left-0 z-20 mt-2 w-36 origin-top-right divide-y divide-gray-100 rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none"
                   style={{ marginLeft: '14rem' }}
                 >
-                  <div
-                    role="button"
-                    tabIndex={0}
-                    onFocus={() => {
-                      setFocusedElement('subMenuFirstItem');
-                    }}
-                    className="py-1">
-                    <Menu.Item>
-                      {({ active }: { active: boolean }) => (
-                        <div
-                          className={classNames(
-                            focusedElement == 'subMenuFirstItem' && active ? 'bg-blue-400 text-gray-900' : 'text-gray-700',
-                            'block px-4 py-2 text-sm'
-                          )}
-                        >
-                          <div className="flex gap-x-2">
-                            <div>
-                              <SiGithub size="1.3em" color="black" />
-                            </div>
-                            <div onFocus={() => setFocusedElement('subMenuFirstItem')}>Github</div>
-                          </div>
-                        </div>
-                      )}
-                    </Menu.Item>
-                    <Menu.Item>
-                      {({ active }: { active: boolean }) => (
-                        <div
-                          role="button"
-                          tabIndex={0}
-                          onFocus={() => {
-                            setFocusedElement('subMenuSecondItem');
-                          }}
-                          className={classNames(
-                            active ? 'bg-blue-400 text-gray-900' : 'text-gray-700',
-                            'block px-4 py-2 text-sm'
-                          )}
-                        >
-                          <div className="flex gap-x-2">
-                            <a>
-                              <RxStitchesLogo size="1.3em" color="black" />
-                            </a>
-                            <a> Sititches</a>
-                          </div>
-                        </div>
-                      )}
-                    </Menu.Item>
-                    <Menu.Item>
-                      {({ active }: { active: boolean }) => (
-                        <div
-                          onFocus={() => {
-                            setFocusedElement('subMenuThirdItem');
-                          }}
-                          className={classNames(
-                            active ? 'bg-blue-400 text-gray-900' : 'text-gray-700',
-                            'block px-4 py-2 text-sm'
-                          )}
-                        >
-                          <div className="flex gap-x-2">
-                            <a>
-                              <RxTwitterLogo size="1.3em" color="black" />
-                            </a>
-                            <a> Sititches</a>
-                          </div>
-                        </div>
-                      )}
-                    </Menu.Item>
-                  </div>
+                  <DropdownMenuItem label="GitHub" Icon={SiGithub} checked={true} />
+                  <DropdownMenuItem label='Stitches' Icon={RxStitchesLogo} checked={true} />
+                  <DropdownMenuItem label='Twitter' Icon={RxTwitterLogo} checked={true} />
                 </div>
               )}
-              <Menu.Item>
-                {({ active }: { active: boolean }) => (
-                  <a
-                    href="#"
-                    className={classNames(
-                      active ? 'bg-blue-400 text-gray-900' : 'text-gray-700',
-                      'block px-4 py-2 text-sm'
-                    )}
-                  >
-                    <div className='flex flex-row '>
-                      <div className='w-6'>{checked === 'download' ? <BsCheck size="1.5em" color="black" /> : ''} </div>
-                      <div onFocus={() => setFocusedElement('download')}>Download</div>
-                    </div>
-                  </a>
-                )}
-              </Menu.Item>
+              <DropdownMenuItem label='Downloads' Icon={BsCheck} checked={checked === 'downloads'} />
             </div>
             <div className="py-1">
-              <Menu.Item>
-                {({ active }: { active: boolean }) => (
-                  <div
-                    className={classNames(
-                      active ? 'bg-blue-400 text-gray-900' : 'text-gray-700',
-                      'block px-4 py-2 text-sm'
-                    )}
-                  >
-                    <div className='flex flex-row '>
-                      <div className='w-6'>{checked === 'showToolbar' ? <BsCheck size="1.5em" color="black" /> : ''} </div>
-                      <div>Show Toolbar</div>
-                    </div>
-                  </div>
-                )}
-              </Menu.Item>
-              <Menu.Item>
-                {({ active }: { active: boolean }) => (
-                  <div
-                    className={classNames(
-                      active ? 'bg-blue-400 text-gray-900' : 'text-gray-700',
-                      'block px-4 py-2 text-sm'
-                    )}
-                  >
-                    <div className='flex flex-row '>
-                      <div className='w-6'>{checked === 'showFullUrl' ? <BsCheck size="1.5em" color="black" /> : ''} </div>
-                      <div>Show Full URLs</div>
-                    </div>
-                  </div>
-                )}
-              </Menu.Item>
+              <DropdownMenuItem label='Show Toolbar' Icon={BsCheck} checked={checked === 'showToolbar'} />
+              <DropdownMenuItem label='Show Full URLs' Icon={BsCheck} checked={checked === 'showFullURLs'} />
             </div>
             <div className="py-1">
-              <Menu.Item>
-                {({ active }: { active: boolean }) => (
-                  <div
-                    className={classNames(
-                      active ? 'bg-blue-400 text-gray-900' : 'text-gray-700',
-                      'block px-4 py-2 text-sm'
-                    )}
-                  >
-                    <div className='flex flex-row '>
-                      <div className='w-6'>{checked === 'delete' ? <BsCheck size="1.5em" color="black" /> : ''} </div>
-                      <div>Delete</div>
-                    </div>
-                  </div>
-                )}
-              </Menu.Item>
+              <DropdownMenuItem label='Delete' Icon={BsCheck} checked={checked === 'delete'} />
             </div>
           </Menu.Items>
         </Transition>
